refactor(ingest): split embedding providers into per-file modules
Promotes embedding_provider.py to an embedding_providers/ package with
one file per provider (base, bedrock, cohere, openai, vertex_ai) plus a
factory module. Public API is re-exported from the package __init__ and
unchanged for callers.
